# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant Platform API
  version: 1.0.0
  description: >
    REST API para el sistema de gestión de pedidos, reseñas y analítica de
    restaurantes sobre MongoDB Atlas (CC3089 Bases de Datos 2).


    **Recursos principales**: Users, Restaurants, Menu Items, Carts, Orders,
    Reviews, Files (GridFS), Analytics.
servers:
  - url: http://localhost:3000
    description: Servidor local de desarrollo
...
paths:
...
  /carts:
    get:
      summary: Obtener carrito del usuario
      tags: [Carts]
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: restaurantId
          schema:
            type: string
      responses:
        '200':
          description: Carrito encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: userId requerido.
        '404':
          description: Carrito no encontrado.

    delete:
      summary: Eliminar carrito completo
      tags: [Carts]
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: restaurantId
          schema:
            type: string
      responses:
        '200':
          description: Carrito eliminado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResult'
  /carts/items:
    post:
      summary: Agregar ítem al carrito (upsert + $push + $setOnInsert)
      description: >
        Si el carrito no existe, se crea automáticamente (upsert).
        El ítem se valida contra menu_items y se agrega con embedded snapshot.
      tags: [Carts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, menuItemId]
              properties:
                userId:
                  type: string
                menuItemId:
                  type: string
                quantity:
                  type: integer
                  default: 1
      responses:
        '200':
          description: Carrito actualizado con el nuevo ítem.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '404':
          description: Ítem de menú no encontrado o no disponible.
  /carts/items/{menuItemId}:
    patch:
      summary: Actualizar cantidad de ítem en carrito (positional $ operator)
      tags: [Carts]
      parameters:
        - in: path
          name: menuItemId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, quantity]
              properties:
                userId:
                  type: string
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Carrito actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: quantity debe ser >= 1.
        '404':
          description: Carrito o ítem no encontrado.

    delete:
      summary: Eliminar ítem del carrito ($pull)
      tags: [Carts]
      parameters:
        - in: path
          name: menuItemId
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ítem eliminado del carrito.
          content:
            application/json:
              schema:
                type: object
                properties:
                  removed:
                    type: integer
...
  /analytics/count:
    get:
      summary: Contar documentos (countDocuments — agregación simple)
      tags: [Analytics]
      parameters:
        - in: query
          name: collection
          required: true
          schema:
            type: string
          description: Nombre de la colección (orders, reviews, menu_items, etc.).
        - in: query
          name: restaurantId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: Conteo de documentos.
          content:
            application/json:
              schema:
                type: object
                properties:
                  collection:
                    type: string
                  count:
                    type: integer
  /analytics/distinct:
    get:
      summary: Valores distintos de un campo (distinct — agregación simple)
      tags: [Analytics]
      parameters:
        - in: query
          name: collection
          required: true
          schema:
            type: string
        - in: query
          name: field
          required: true
          schema:
            type: string
          description: Campo sobre el que obtener valores distintos.
        - in: query
          name: restaurantId
          schema:
            type: string
      responses:
        '200':
          description: Valores distintos.
          content:
            application/json:
              schema:
                type: object
                properties:
                  collection:
                    type: string
                  field:
                    type: string
                  values:
                    type: array
                    items: {}
  /analytics/top-restaurants:
    get:
      summary: "Pipeline 1: Restaurantes mejor calificados ($group + $match + $lookup)"
      tags: [Analytics]
      parameters:
        - in: query
          name: minReviews
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Ranking de restaurantes por rating promedio.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    restaurant:
                      type: object
                      properties:
                        name:
                          type: string
                        cuisineTypes:
                          type: array
                          items:
                            type: string
                    avgRating:
                      type: number
                    reviewCount:
                      type: integer
  /analytics/best-selling-items:
    get:
      summary: "Pipeline 2: Ítems más vendidos ($unwind + $group + $lookup)"
      tags: [Analytics]
      parameters:
        - in: query
          name: restaurantId
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Ranking de ítems por cantidad vendida.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    totalQty:
                      type: integer
                    totalRevenue:
                      type: number
                    category:
                      type: string
  /analytics/revenue-by-month:
    get:
      summary: "Pipeline 3: Ingresos por mes ($group temporal + $lookup)"
      tags: [Analytics]
      parameters:
        - in: query
          name: restaurantId
          schema:
            type: string
      responses:
        '200':
          description: Ingresos agrupados por restaurante-mes.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    restaurantName:
                      type: string
                    year:
                      type: integer
                    month:
                      type: integer
                    totalRevenue:
                      type: number
                    orderCount:
                      type: integer
                    avgOrderValue:
                      type: number
  /analytics/rating-distribution/{restaurantId}:
    get:
      summary: "Pipeline 4: Distribución de ratings ($group + $push)"
      tags: [Analytics]
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Distribución de ratings 1-5.
          content:
            application/json:
              schema:
                type: object
                properties:
                  distribution:
                    type: array
                    items:
                      type: object
                      properties:
                        rating:
                          type: integer
                        count:
                          type: integer
                  total:
                    type: integer
  /analytics/order-velocity/{restaurantId}:
    get:
      summary: "Pipeline 5: Velocidad de pedidos ($dateTrunc ventanas 5 min — Time Series)"
      tags: [Analytics]
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema:
            type: string
        - in: query
          name: hoursBack
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Conteo de pedidos por ventana temporal de 5 minutos.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    window:
                      type: string
                      format: date-time
                    count:
                      type: integer
  /analytics/avg-transition-time/{restaurantId}:
    get:
      summary: "Pipeline 6: Tiempo promedio entre transiciones de estado"
      tags: [Analytics]
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Duración promedio por transición.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    from:
                      type: string
                    to:
                      type: string
                    avgDurationSec:
                      type: number
                    count:
                      type: integer
  /analytics/tags:
    get:
      summary: Tags más usados en reseñas ($unwind + $group)
      tags: [Analytics]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 15
      responses:
        '200':
          description: Ranking de tags.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    tag:
                      type: string
                    count:
                      type: integer
...
  /analytics/revenue-by-category:
    get:
      summary: Ingresos por categoría ($unwind items + $lookup menu_items)
      tags: [Analytics]
      parameters:
        - in: query
          name: restaurantId
          schema:
            type: string
      responses:
        '200':
          description: Ingresos agrupados por categoría de menú.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    category:
                      type: string
                    totalRevenue:
                      type: number
                    totalQty:
                      type: integer
  /analytics/daily-revenue:
    get:
      summary: Ingresos diarios pre-computados
      tags: [Analytics]
      parameters:
        - in: query
          name: restaurantId
          schema:
            type: string
        - in: query
          name: days
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Registros diarios de ingresos.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    restaurantId:
                      type: string
                    date:
                      type: string
                      format: date-time
                    revenue:
                      type: number
                    orderCount:
                      type: integer
  /analytics/restaurant-stats:
    get:
      summary: Estadísticas pre-computadas de restaurantes (materialized view)
      tags: [Analytics]
      parameters:
        - $ref: '#/components/parameters/skip'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Stats por restaurante.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    restaurantName:
                      type: string
                    totalOrders:
                      type: integer
                    totalDelivered:
                      type: integer
                    totalCancelled:
                      type: integer
                    totalRevenue:
                      type: number
                    avgOrderValue:
                      type: number
                    totalReviews:
                      type: integer
                    avgRating:
                      type: number
  /analytics/run-batch:
    post:
      summary: Ejecutar batch job manualmente ($merge materialized views)
      tags: [Analytics]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                job:
                  type: string
                  enum: [daily, weekly]
                  default: daily
                targetDate:
                  type: string
                  format: date-time
                  description: Fecha objetivo (solo para job=daily).
      responses:
        '200':
          description: Job ejecutado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  job:
                    type: string
                  count:
                    type: integer
...
components:
  parameters:
...
    skip:
      in: query
      name: skip
      schema:
        type: integer
        minimum: 0
        default: 0
    limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Cart:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        restaurantId:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: string
              name:
                type: string
              price:
                type: number
              quantity:
                type: integer
              subtotal:
                type: number
              available:
                type: boolean
        subtotal:
          type: number
        hasUnavailableItems:
          type: boolean
        expiresAt:
          type: string
          format: date-time
...
    DeleteResult:
      type: object
      properties:
        deleted:
          type: integer
          example: 1
...